refactor(store): Move trait impls to separate modules
