+++ rust/src/store/canonicity_store_impl.rs
@@ -1,0 +1,184 @@
+use super::{column_families::ColumnFamilyHelpers, fixed_keys::FixedKeys};
+use crate::{
+    block::{precomputed::PrecomputedBlock, store::BlockStore, BlockHash},
+    canonicity::{store::CanonicityStore, Canonicity},
+    event::{db::*, store::EventStore, IndexerEvent},
+    snark_work::store::SnarkStore,
+    store::IndexerStore,
+};
+use log::trace;
+
+impl CanonicityStore for IndexerStore {
+    fn add_canonical_block(
+        &self,
+        height: u32,
+        state_hash: &BlockHash,
+        genesis_state_hash: &BlockHash,
+        genesis_prev_state_hash: Option<&BlockHash>,
+    ) -> anyhow::Result<()> {
+        if state_hash == genesis_state_hash && genesis_prev_state_hash.is_some() {
+            trace!("Adding new genesis block (length {height}): {state_hash}");
+        } else {
+            trace!("Adding canonical block (length {height}): {state_hash}");
+        }
+
+        // height -> state hash
+        self.database.put_cf(
+            self.canonicity_cf(),
+            height.to_be_bytes(),
+            serde_json::to_vec(state_hash)?,
+        )?;
+
+        // update canonical chain length
+        self.set_max_canonical_blockchain_length(height)?;
+
+        // update top snarkers based on the incoming canonical block
+        if let Some(completed_works) = self.get_snark_work_in_block(state_hash)? {
+            self.update_top_snarkers(completed_works)?;
+        }
+
+        // record new genesis/prev state hashes
+        if let Some(genesis_prev_state_hash) = genesis_prev_state_hash {
+            let (mut genesis_state_hashes, mut genesis_prev_state_hashes) = (
+                self.get_known_genesis_state_hashes()?,
+                self.get_known_genesis_prev_state_hashes()?,
+            );
+
+            // check if genesis hash is present
+            if !genesis_state_hashes.contains(genesis_state_hash) {
+                // if not
+                // add genesis state hash
+                genesis_state_hashes.push(genesis_state_hash.clone());
+                self.database.put(
+                    Self::KNOWN_GENESIS_STATE_HASHES_KEY,
+                    serde_json::to_vec(&genesis_state_hashes)?,
+                )?;
+
+                // add genesis prev state hash
+                genesis_prev_state_hashes.push(genesis_prev_state_hash.clone());
+                self.database.put(
+                    Self::KNOWN_GENESIS_PREV_STATE_HASHES_KEY,
+                    serde_json::to_vec(&genesis_prev_state_hashes)?,
+                )?;
+            }
+        }
+
+        // record new canonical block event
+        self.add_event(&IndexerEvent::Db(DbEvent::Canonicity(
+            DbCanonicityEvent::NewCanonicalBlock {
+                blockchain_length: height,
+                state_hash: state_hash.0.clone().into(),
+            },
+        )))?;
+        Ok(())
+    }
+
+    fn get_known_genesis_state_hashes(&self) -> anyhow::Result<Vec<BlockHash>> {
+        trace!("Getting known genesis state hashes");
+        Ok(self
+            .database
+            .get(Self::KNOWN_GENESIS_STATE_HASHES_KEY)?
+            .map_or(vec![], |bytes| {
+                serde_json::from_slice(&bytes).expect("known genesis state hashes")
+            }))
+    }
+
+    fn get_known_genesis_prev_state_hashes(&self) -> anyhow::Result<Vec<BlockHash>> {
+        trace!("Getting known genesis prev state hashes");
+        Ok(self
+            .database
+            .get(Self::KNOWN_GENESIS_PREV_STATE_HASHES_KEY)?
+            .map_or(vec![], |bytes| {
+                serde_json::from_slice(&bytes).expect("known genesis prev state hashes")
+            }))
+    }
+
+    fn get_canonical_hash_at_height(&self, height: u32) -> anyhow::Result<Option<BlockHash>> {
+        trace!("Getting canonical hash at height {height}");
+
+        let key = height.to_be_bytes();
+        let canonicity_cf = self.canonicity_cf();
+        match self
+            .database
+            .get_pinned_cf(&canonicity_cf, key)?
+            .map(|bytes| bytes.to_vec())
+        {
+            None => Ok(None),
+            Some(bytes) => Ok(Some(serde_json::from_slice(&bytes)?)),
+        }
+    }
+
+    fn get_max_canonical_blockchain_length(&self) -> anyhow::Result<Option<u32>> {
+        trace!("Getting max canonical blockchain length");
+
+        let canonicity_cf = self.canonicity_cf();
+        match self
+            .database
+            .get_pinned_cf(&canonicity_cf, Self::MAX_CANONICAL_KEY)?
+            .map(|bytes| bytes.to_vec())
+        {
+            None => Ok(None),
+            Some(bytes) => Ok(Some(serde_json::from_slice(&bytes)?)),
+        }
+    }
+
+    fn set_max_canonical_blockchain_length(&self, height: u32) -> anyhow::Result<()> {
+        trace!("Setting max canonical blockchain length to {height}");
+
+        let canonicity_cf = self.canonicity_cf();
+        let value = serde_json::to_vec(&height)?;
+        self.database
+            .put_cf(&canonicity_cf, Self::MAX_CANONICAL_KEY, value)?;
+        Ok(())
+    }
+
+    fn get_block_canonicity(&self, state_hash: &BlockHash) -> anyhow::Result<Option<Canonicity>> {
+        trace!("Getting canonicity of block with hash {}", state_hash.0);
+
+        if let Ok(Some(best_tip)) = self.get_best_block() {
+            if let Some(blockchain_length) = self.get_block(state_hash)?.map(|pcb| match pcb {
+                PrecomputedBlock::V1(v1) => v1.blockchain_length,
+                PrecomputedBlock::V2(pcb_v2) => {
+                    pcb_v2.protocol_state.body.consensus_state.blockchain_length
+                }
+            }) {
+                if blockchain_length > best_tip.blockchain_length() {
+                    return Ok(None);
+                } else if let Some(max_canonical_length) =
+                    self.get_max_canonical_blockchain_length()?
+                {
+                    if blockchain_length > max_canonical_length {
+                        // follow best chain back from tip to given block
+                        let mut curr_block = best_tip;
+                        while curr_block.state_hash() != *state_hash
+                            && curr_block.blockchain_length() > max_canonical_length
+                        {
+                            if let Some(parent) =
+                                self.get_block(&curr_block.previous_state_hash())?
+                            {
+                                curr_block = parent;
+                            } else {
+                                break;
+                            }
+                        }
+
+                        if curr_block.state_hash() == *state_hash
+                            && curr_block.blockchain_length() > max_canonical_length
+                        {
+                            return Ok(Some(Canonicity::Canonical));
+                        } else {
+                            return Ok(Some(Canonicity::Orphaned));
+                        }
+                    } else if self.get_canonical_hash_at_height(blockchain_length)?
+                        == Some(state_hash.clone())
+                    {
+                        return Ok(Some(Canonicity::Canonical));
+                    } else {
+                        return Ok(Some(Canonicity::Orphaned));
+                    }
+                }
+            }
+        }
+        Ok(None)
+    }
+}
